Add class RappersReader that creates a rappers array
